chore: remove old recents page and rename home
