+++ packages/app/src/app/overmind/namespaces/dashboard/types.ts
@@ -2,8 +2,7 @@ import { SidebarCollectionDashboardFragment as Collection } from 'app/graphql/ty
 
 export type PageTypes =
   | 'search'
-  | 'home'
-  | 'recents'
+  | 'recent'
   | 'archive'
   | 'templates'
   | 'drafts'
@@ -35,9 +34,7 @@ export enum sandboxesTypes {
   RECENT = 'RECENT',
   SHARED = 'SHARED',
   LIKED = 'LIKED',
-  HOME = 'HOME',
   TEMPLATE_HOME = 'TEMPLATE_HOME',
-  RECENT_HOME = 'RECENT_HOME',
   ALL = 'ALL',
   SEARCH = 'SEARCH',
   REPOS = 'REPOS',
