+++ packages/app/src/app/pages/Dashboard/Content/routes/Recent/index.tsx
@@ -1,77 +1,111 @@
-import React from 'react';
-import { Helmet } from 'react-helmet';
-import { SelectionProvider } from 'app/pages/Dashboard/Components/Selection';
+import React, { useEffect } from 'react';
 import { useAppState, useActions } from 'app/overmind';
 import { sandboxesTypes } from 'app/overmind/namespaces/dashboard/types';
 import { Header } from 'app/pages/Dashboard/Components/Header';
-import { VariableGrid } from 'app/pages/Dashboard/Components/VariableGrid';
+import {
+  VariableGrid,
+  GUTTER,
+  GRID_MAX_WIDTH,
+} from 'app/pages/Dashboard/Components/VariableGrid';
+import { SelectionProvider } from 'app/pages/Dashboard/Components/Selection';
+import { Helmet } from 'react-helmet';
 import {
   DashboardGridItem,
-  DashboardHeader,
-  DashboardSandbox,
+  DashboardTemplate,
   PageTypes,
 } from 'app/pages/Dashboard/types';
-import { getPossibleTemplates } from '../../utils';
+import styled from 'styled-components';
+
+const BannerProjects = styled.a`
+  display: block;
+
+  width: calc(100% - ${2 * GUTTER}px);
+  max-width: ${GRID_MAX_WIDTH - 2 * GUTTER}px;
+  margin: 0 auto ${GUTTER}px;
+
+  img {
+    width: 100%;
+  }
+`;
 
 export const Recent = () => {
   const {
     activeTeam,
-    dashboard: { sandboxes, recentSandboxesByTime, getFilteredSandboxes },
+    dashboard: { viewMode, sandboxes },
   } = useAppState();
   const {
     dashboard: { getPage },
   } = useActions();
 
-  React.useEffect(() => {
+  useEffect(() => {
     getPage(sandboxesTypes.RECENT);
-  }, [activeTeam, getPage]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [activeTeam]);
 
-  const getSection = (
-    title: string,
-    time: keyof typeof recentSandboxesByTime
-  ): [DashboardHeader, ...DashboardSandbox[]] | [] => {
-    const recentSandboxes = getFilteredSandboxes(recentSandboxesByTime[time]);
+  const templates: DashboardGridItem[] = (sandboxes.TEMPLATE_HOME || []).map(
+    (template, i) => {
+      const { sandbox, ...templateValues } = template;
 
-    if (!recentSandboxes.length) return [];
-
-    return [
-      { type: 'header', title },
-      ...recentSandboxes.map(sandbox => ({
-        type: 'sandbox' as 'sandbox',
+      const dashboardTemplate: DashboardTemplate = {
+        type: 'template' as 'template',
         sandbox,
-      })),
-    ];
-  };
+        template: templateValues,
+        autoFork: true,
+        noDrag: true,
+        optional: i >= 3,
+      };
+
+      return dashboardTemplate;
+    }
+  );
+
+  // Make sure that a new row starts for the next item
+  templates.push({ type: 'blank-row-fill' });
 
   const items: DashboardGridItem[] = sandboxes.RECENT
     ? [
-        ...getSection('Today', 'day'),
-        ...getSection('Last 7 days', 'week'),
-        ...getSection('Earlier this month', 'month'),
-        ...getSection('Older', 'older'),
+        {
+          type: 'header',
+          title: 'Recently Used Templates',
+          ...(viewMode === 'list'
+            ? {
+                showMoreLink: '/s',
+                showMoreLabel: '+ New Sandbox',
+              }
+            : {}),
+        },
+        { type: 'new-sandbox' },
+        ...templates,
+        {
+          type: 'header',
+          title: 'Recently Viewed Sandboxes',
+        },
+        ...(sandboxes.RECENT || []).map(sandbox => ({
+          type: 'sandbox' as 'sandbox',
+          sandbox,
+        })),
       ]
     : [
-        { type: 'header', title: 'Today' },
+        { type: 'header', title: 'Recently Used Templates' },
         { type: 'skeleton-row' },
-        { type: 'header', title: 'Last 7 days' },
+        { type: 'header', title: 'Recently Viewed Sandboxes' },
         { type: 'skeleton-row' },
       ];
 
-  const pageType: PageTypes = 'recents';
+  const pageType: PageTypes = 'recent';
   return (
-    <SelectionProvider page={pageType} activeTeamId={activeTeam} items={items}>
+    <SelectionProvider activeTeamId={activeTeam} page={pageType} items={items}>
       <Helmet>
-        <title>Recent Sandboxes - CodeSandbox</title>
+        <title>Dashboard - CodeSandbox</title>
       </Helmet>
-      <Header
-        title="Recently Modified Sandboxes"
-        activeTeam={activeTeam}
-        templates={getPossibleTemplates(sandboxes.RECENT)}
-        showViewOptions
-        showFilters
-      />
-
-      <VariableGrid items={items} page={pageType} />
+      <BannerProjects href="https://codesandbox.io/p/dashboard?from=banner">
+        <img
+          src="/static/img/projects-banner.png"
+          alt="The CodeSandbox flow, for any project of any size. Try now"
+        />
+      </BannerProjects>
+      <Header title="Recent" activeTeam={activeTeam} showViewOptions />
+      <VariableGrid page={pageType} items={items} />
     </SelectionProvider>
   );
 };
